[ROUTES] restructure repo to put routes in a separate file
